Analyzer -> Analysis rename and related cleanup
