--- lib/how_is/analysis.rb
@@ -1,8 +1,0 @@
-require "ostruct"
-
-class HowIs
-  ##
-  # Represents a completed analysis of the repository being analyzed.
-  class Analysis < OpenStruct
-  end
-end
