+++ spec/how_is/analyzer_spec.rb
@@ -2,7 +2,7 @@
 
 require "spec_helper"
 
-describe HowIs::Analyzer do
+describe HowIs::Analysis do
   let(:issues) { JSON.parse(open(File.expand_path("../data/issues.json", __dir__)).read) }
   let(:pulls) { JSON.parse(open(File.expand_path("../data/pulls.json", __dir__)).read) }
 
@@ -11,7 +11,7 @@ describe HowIs::Analyzer do
 
   let(:fetcher_results) { HowIs::Fetcher::Results.new(issues, pulls) }
 
-  subject { HowIs::Analyzer.new }
+  subject { HowIs::Analysis.new }
 
   context "#num_with_label" do
     it "returns a Hash mapping labels to the number of issues or pulls with that label" do
